requirements.txt
aio-pika==9.3.0
aiofiles==23.2.1
aiogram==2.19
aiohttp==3.8.6
aiormq==6.7.7
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.1.0
Babel==2.9.1
beautifulsoup4==4.10.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==36.0.1
dnspython==2.4.2
frozenlist==1.4.0
idna==3.4
motor==3.3.2
msgpack==1.0.7
multidict==6.0.4
pamqp==3.2.1
psycopg2-binary==2.9.9
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pymongo==4.6.1
python-dotenv==1.0.0
pytz==2023.3.post1
soupsieve==2.5
SQLAlchemy==2.0.23
typing_extensions==4.8.0
yarl==1.9.2
